v0.5.4-rc0 - Sample rate / Pos X / Offset / Period: drop QDoubleValidator that   blocked SI suffix keystrokes (k, M, G, m, u, n). All physical-unit   fields now parse via parseSIValue uniformly (e.g. "10MHz", "125u"). - formatSIValueSigned: add m/u/n/p prefixes so sub-1 values keep   precision. Symbol period read-out previously collapsed to "0.000s"   for typical us..ms periods; now displays "125.0us" etc.

+++ src/abstractsamplesource.cpp
@@ -33,7 +33,9 @@ void AbstractSampleSource::invalidate()
 
 int AbstractSampleSource::subscriberCount()
 {
-    return subscribers.size();
+    /* explicit cast: subscriber count is bounded by the number of
+     * plot/UI consumers (single-digit), int is more than enough */
+    return (int)subscribers.size();
 }
 
 void AbstractSampleSource::unsubscribe(Subscriber *subscriber)
